+++ src/Suave.IIS/Configuration.fs
@@ -5,13 +5,14 @@ open System.Net
 
 type Setup = {
     Port: uint16
-    Path: string
+    Path: string option
 }
 
 /// Parse configuration setup from command line args
 let parseSetup (args:string []) =
     match args |> Array.length with
-    | x when x >= 2 -> { Port = uint16 args.[0]; Path = args.[1] } |> Some
+    | x when x = 2 -> { Port = uint16 args.[0]; Path = Some args.[1] } |> Some
+    | x when x = 1 -> { Port = uint16 args.[0]; Path = None } |> Some
     | _  -> None
 
 /// Set Suave port based on command line args setup